Refactor RatingValue to use double type
- Updated the database migration to change the RatingValue column from string to double in the Ratings table.
- Modified CreateRatingDto to reflect the new double type for RatingValue.
- Adjusted RatingDto to ensure consistency with the updated RatingValue type.

+++ Tradof.CompanyModule.Services/DTOs/CompanyDTOs.cs
@@ -11,7 +11,7 @@ namespace Tradof.CompanyModule.Services.DTOs
         string Email,
         string Phone,
         string JobTitle,
-        int RatingSum,
+        double RatingSum,
         int ReviewCount,
         string ProfileImageUrl,
         long CountryId,
